Pushing fully functional download script and deleted all non-functional ones.
