+++ bird_data/Europe_extended_02.csv
@@ -29,3 +29,50 @@ species_no,scientific_name
 8042,loxia scotica
 10230,lullula arborea
 10917,luscinia megarhynchos
+10917,luscinia megarhynchos
+10775,mergus merganser
+9861,oenanthe hispanica
+6134,oenanthe xanthoprymna
+7879,onychoprion aleuticus
+8953,passer italiae
+10904,perdix perdix
+10700,pernis apivorus
+7477,phoenicurus moussieri
+9260,phylloscopus bonelli
+10701,phylloscopus collybita
+8692,phylloscopus ibericus
+7379,phylloscopus orientalis
+10209,picoides tridactylus
+5718,picus sharpei
+6406,picus vaillantii
+10840,picus viridis
+10633,pinicola enucleator
+10538,podiceps auritus
+10471,podiceps grisegena
+8396,poecile lugubris
+10667,prunella modularis
+9384,puffinus mauretanicus
+9089,puffinus yelkouan
+9134,pyrrhula murina
+10459,regulus ignicapilla
+10454,remiz pendulinus
+10568,saxicola rubetra
+10961,serinus canaria
+10654,serinus serinus
+6249,sitta krueperi
+6840,sitta ledanti
+8826,sitta neumayer
+7705,sitta whiteheadi
+10937,strix aluco
+10907,strix nebulosa
+10100,sturnus unicolor
+5355,sylvia balearica
+10457,sylvia borin
+10436,sylvia communis
+6563,sylvia deserticola
+9481,sylvia hortensis
+7325,sylvia sarda
+9974,sylvia undata
+9642,synthliboramphus antiquus
+8619,tetraogallus caucasicus
+10423,turdus torquatus
