+++ bird_data/test_2_tracking.csv
@@ -1,0 +1,44 @@
+species_no,downloaded,downloaded_no
+1,1,10
+2,1,15
+3,1,17
+1,1,5
+1,1,1
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,4
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
+1,1,5
+2,1,7
+3,1,5
