--- bird_data/test_tracking.csv
@@ -1,4 +1,0 @@
-species_no,downloaded,downloaded_no
-1,1,5
-2,1,7
-3,1,5
