+++ bird_data/tracking.csv
@@ -96,3 +96,40 @@ species_no,downloaded,downloaded_no
 10100,1,435
 5355,1,239
 10457,1,427
+10172,1,425
+10951,1,455
+10220,1,365
+10024,1,411
+10560,1,431
+10279,1,428
+8029,1,352
+10569,1,368
+10909,1,409
+10087,1,431
+7618,1,392
+9616,1,419
+10734,1,365
+9980,1,385
+9223,1,396
+10332,1,407
+10759,1,436
+8353,1,433
+10854,1,430
+10398,1,169
+10851,1,540
+9149,1,437
+9874,1,434
+9863,1,456
+10659,1,395
+10267,1,412
+10194,1,419
+8042,1,297
+10230,1,433
+10436,1,430
+6563,1,227
+9481,1,418
+7325,1,347
+9974,1,422
+9642,1,413
+8619,1,308
+10423,1,427
